requirements.txt
# ── Core Framework ──────────────────────────────────────────────────
flask>=3.0
gunicorn>=21.0

# ── Data Processing & ML ────────────────────────────────────────────
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
numpy==1.26.4

# ── Image Duplicate Detection ──────────────────────────────────────
imagehash>=4.3
Pillow>=10.0

# ── Image Classification (optional — adds ~2GB, skip for free tier) ─
# torch>=2.0
# ultralytics>=8.0
# opencv-python
